Moved isCheckmateCompatibleWithGame() from wincondition.ts -> winconutil.ts
