+++ src/shared/chess/logic/wincondition.ts
@@ -18,11 +18,7 @@ import typeutil, { RawType } from '../util/typeutil.js';
 import { detectRepetitionDraw } from './repetition.js';
 import { rawTypes as r, Player } from '../util/typeutil.js';
 import { detectInsufficientMaterial } from './insufficientmaterial.js';
-import {
-	detectCheckmateOrStalemate,
-	pieceCountToDisableCheckmate,
-	royalCountToDisableCheckmate,
-} from './checkmate.js';
+import { detectCheckmateOrStalemate } from './checkmate.js';
 
 // The squares in KOTH where if you get your king to you WIN
 // prettier-ignore
@@ -192,27 +188,6 @@ function wasLastMoveARoyalCapture(boardsim: Board): boolean | undefined {
 	return false;
 }
 
-/**
- * If the game is multiplayer, or if anyone gets multiple turns in a row, then that allows capturing
- * of the kings no matter the win conditions, by way of one person opening a discovered on turn 1, and
- * another person capturing the king on turn 2 => CHECKMATE NOT COMPATIBLE!
- *
- * Checkmate is also not compatible with games with colinear lines present, because the logic surrounding
- * making opening discovered attacks illegal is a nightmare.
- * @param gamefile
- * @returns true if the gamefile is checkmate compatible
- */
-function isCheckmateCompatibleWithGame({ boardsim, basegame }: FullGame): boolean {
-	if (boardsim.editor) return false; // This prevents legal move calculation respecting check in the editor.
-	if (boardutil.getPieceCountOfGame(boardsim.pieces) > pieceCountToDisableCheckmate) return false; // Too many pieces (checkmate algorithm takes too long)
-	if (boardsim.pieces.slides.length > 16) return false; // If the game has more lines than this, then checkmate creates lag spikes.
-	if (gamefileutility.getPlayerCount(basegame) > 2) return false; // 3+ Players allows for 1 player to open a discovered and a 2nd to capture a king. CHECKMATE NOT COMPATIBLE
-	if (moveutil.doesAnyPlayerGet2TurnsInARow(basegame)) return false; // This also allows the capture of the king.
-	if (boardutil.getRoyalCountOfGame(boardsim.pieces) > royalCountToDisableCheckmate) return false; // Too many royals (check & checkmate algorithm takes too long)
-	return true; // Checkmate compatible!
-}
-
 export default {
 	getGameConclusion,
-	isCheckmateCompatibleWithGame,
 };
