+++ src/shared/chess/util/winconutil.ts
@@ -5,11 +5,16 @@
  * And contains a few utility methods for them.
  */
 
+import type { FullGame } from '../logic/fullgame.js';
 import type { GameRules } from './gamerules.js';
 
 import * as z from 'zod';
 
+import moveutil from './moveutil.js';
+import boardutil from './boardutil.js';
 import typeschemas from './typeschemas.js';
+import gamefileutility from './gamefileutility.js';
+import { pieceCountToDisableCheckmate, royalCountToDisableCheckmate } from '../logic/checkmate.js';
 
 // Constants -----------------------------------------------------------------
 
@@ -117,6 +122,26 @@ const TERMINATION_IN_ENGLISH = {
 // Functions --------------------------------------------------------------------------
 
 /**
+ * If the game is multiplayer, or if anyone gets multiple turns in a row, then that allows capturing
+ * of the kings no matter the win conditions, by way of one person opening a discovered on turn 1, and
+ * another person capturing the king on turn 2 => CHECKMATE NOT COMPATIBLE!
+ *
+ * Checkmate is also not compatible with games with colinear lines present, because the logic surrounding
+ * making opening discovered attacks illegal is a nightmare.
+ * @param gamefile
+ * @returns true if the gamefile is checkmate compatible
+ */
+function isCheckmateCompatibleWithGame({ boardsim, basegame }: FullGame): boolean {
+	if (boardsim.editor) return false; // This prevents legal move calculation respecting check in the editor.
+	if (boardutil.getPieceCountOfGame(boardsim.pieces) > pieceCountToDisableCheckmate) return false; // Too many pieces (checkmate algorithm takes too long)
+	if (boardsim.pieces.slides.length > 16) return false; // If the game has more lines than this, then checkmate creates lag spikes.
+	if (gamefileutility.getPlayerCount(basegame) > 2) return false; // 3+ Players allows for 1 player to open a discovered and a 2nd to capture a king. CHECKMATE NOT COMPATIBLE
+	if (moveutil.doesAnyPlayerGet2TurnsInARow(basegame)) return false; // This also allows the capture of the king.
+	if (boardutil.getRoyalCountOfGame(boardsim.pieces) > royalCountToDisableCheckmate) return false; // Too many royals (check & checkmate algorithm takes too long)
+	return true; // Checkmate compatible!
+}
+
+/**
  * Calculates if the provided condition is move-triggered.
  * This is any conclusion that can happen after a move is made.
  * Excludes conclusions like resignation, time, aborted, disconnect,
@@ -147,6 +172,7 @@ export default {
 
 	GAMERULE_WIN_CONDITIONS,
 
+	isCheckmateCompatibleWithGame,
 	isConclusionMoveTriggered,
 	getTerminationInEnglish,
 };
